fix: check for empty array as well

+++ apiserver/plane/api/serializers/issue.py
@@ -100,7 +100,7 @@ class IssueCreateSerializer(BaseSerializer):
         project = self.context["project"]
         issue = Issue.objects.create(**validated_data, project=project)
 
-        if blockers is not None:
+        if blockers is not None and len(blockers):
             IssueBlocker.objects.bulk_create(
                 [
                     IssueBlocker(
@@ -116,7 +116,7 @@ class IssueCreateSerializer(BaseSerializer):
                 batch_size=10,
             )
 
-        if assignees is not None:
+        if assignees is not None and len(assignees):
             IssueAssignee.objects.bulk_create(
                 [
                     IssueAssignee(
@@ -143,7 +143,7 @@ class IssueCreateSerializer(BaseSerializer):
                     updated_by=issue.updated_by,
                 )
 
-        if labels is not None:
+        if labels is not None and len(labels):
             IssueLabel.objects.bulk_create(
                 [
                     IssueLabel(
@@ -159,7 +159,7 @@ class IssueCreateSerializer(BaseSerializer):
                 batch_size=10,
             )
 
-        if blocks is not None:
+        if blocks is not None and len(blocks):
             IssueBlocker.objects.bulk_create(
                 [
                     IssueBlocker(
@@ -183,7 +183,7 @@ class IssueCreateSerializer(BaseSerializer):
         labels = validated_data.pop("labels_list", None)
         blocks = validated_data.pop("blocks_list", None)
 
-        if blockers is not None:
+        if blockers is not None and len(blockers):
             IssueBlocker.objects.filter(block=instance).delete()
             IssueBlocker.objects.bulk_create(
                 [
@@ -200,7 +200,7 @@ class IssueCreateSerializer(BaseSerializer):
                 batch_size=10,
             )
 
-        if assignees is not None:
+        if assignees is not None and len(assignees):
             IssueAssignee.objects.filter(issue=instance).delete()
             IssueAssignee.objects.bulk_create(
                 [
@@ -217,7 +217,7 @@ class IssueCreateSerializer(BaseSerializer):
                 batch_size=10,
             )
 
-        if labels is not None:
+        if labels is not None and len(labels):
             IssueLabel.objects.filter(issue=instance).delete()
             IssueLabel.objects.bulk_create(
                 [
@@ -234,7 +234,7 @@ class IssueCreateSerializer(BaseSerializer):
                 batch_size=10,
             )
 
-        if blocks is not None:
+        if blocks is not None and len(blocks):
             IssueBlocker.objects.filter(blocked_by=instance).delete()
             IssueBlocker.objects.bulk_create(
                 [
